# [BEAM-9167] Reduce Go SDK metric overhead