--- a/sdks/go/pkg/beam/core/metrics/metrics.go
+++ b/sdks/go/pkg/beam/core/metrics/metrics.go
@@ -49,10 +49,13 @@ package metrics
 import (
 	"context"
 	"fmt"
+	"hash/fnv"
 	"sort"
 	"sync"
+	"sync/atomic"
 	"time"
 
+	"github.com/apache/beam/sdks/go/pkg/beam/core/util/ioutilx"
 	"github.com/apache/beam/sdks/go/pkg/beam/log"
 	"github.com/apache/beam/sdks/go/pkg/beam/model/fnexecution_v1"
 	"github.com/golang/protobuf/ptypes"
@@ -64,22 +67,57 @@ import (
 // using metrics requires the PTransform have a context.Context
 // argument.
 
+// perBundle is a struct that retains per transform countersets.
+// TODO(lostluck): Migrate the exec package to use these to extract
+// metric data for export to runner, rather than the global store.
+type perBundle struct {
+	mu  sync.Mutex
+	css []*ptCounterSet
+}
+
+type nameHash uint64
+
+// ptCounterSet is the internal tracking struct for a single ptransform
+// in a single bundle for all counter types.
+type ptCounterSet struct {
+	// We store the user path access to the cells in metric type segregated
+	// maps. At present, caching the name hash,
+	counters      map[nameHash]*counter
+	distributions map[nameHash]*distribution
+	gauges        map[nameHash]*gauge
+}
+
 type ctxKey string
 
-const bundleKey ctxKey = "beam:bundle"
-const ptransformKey ctxKey = "beam:ptransform"
+const (
+	bundleKey     ctxKey = "beam:bundle"
+	ptransformKey ctxKey = "beam:ptransform"
+	counterSetKey ctxKey = "beam:counterset"
+)
 
 // beamCtx is a caching context for IDs necessary to place metric updates.
-//  Allocating contexts and searching for PTransformIDs for every element
+// Allocating contexts and searching for PTransformIDs for every element
 // is expensive, so we avoid it if possible.
 type beamCtx struct {
 	context.Context
 	bundleID, ptransformID string
+	bs                     *perBundle
+	cs                     *ptCounterSet
 }
 
-// Value lifts the beam contLift the keys value for faster lookups when not available.
+// Value implements context.Value for beamCtx, and lifts the
+// values for metrics keys for value for faster lookups.
 func (ctx *beamCtx) Value(key interface{}) interface{} {
 	switch key {
+	case counterSetKey:
+		if ctx.cs == nil {
+			if cs := ctx.Context.Value(key); cs != nil {
+				ctx.cs = cs.(*ptCounterSet)
+			} else {
+				return nil
+			}
+		}
+		return ctx.cs
 	case bundleKey:
 		if ctx.bundleID == "" {
 			if id := ctx.Context.Value(key); id != nil {
@@ -102,24 +140,30 @@ func (ctx *beamCtx) Value(key interface{}) interface{} {
 	return ctx.Context.Value(key)
 }
 
+func (ctx *beamCtx) String() string {
+	return fmt.Sprintf("beamCtx[%s;%s]", ctx.bundleID, ctx.ptransformID)
+}
+
 // SetBundleID sets the id of the current Bundle.
 func SetBundleID(ctx context.Context, id string) context.Context {
 	// Checking for *beamCtx is an optimization, so we don't dig deeply
 	// for ids if not necessary.
 	if bctx, ok := ctx.(*beamCtx); ok {
-		return &beamCtx{Context: bctx.Context, bundleID: id, ptransformID: bctx.ptransformID}
+		return &beamCtx{Context: bctx.Context, bundleID: id, bs: &perBundle{}, ptransformID: bctx.ptransformID}
 	}
-	return &beamCtx{Context: ctx, bundleID: id}
+	return &beamCtx{Context: ctx, bundleID: id, bs: &perBundle{}}
 }
 
 // SetPTransformID sets the id of the current PTransform.
+// Must only be called on a context returened by SetBundleID.
 func SetPTransformID(ctx context.Context, id string) context.Context {
 	// Checking for *beamCtx is an optimization, so we don't dig deeply
 	// for ids if not necessary.
 	if bctx, ok := ctx.(*beamCtx); ok {
-		return &beamCtx{Context: bctx.Context, bundleID: bctx.bundleID, ptransformID: id}
+		return &beamCtx{Context: bctx.Context, bundleID: bctx.bundleID, bs: bctx.bs, ptransformID: id}
 	}
-	return &beamCtx{Context: ctx, ptransformID: id}
+	panic(fmt.Sprintf("SetPTransformID called before SetBundleID for %v", id))
+	return nil // never runs.
 }
 
 const (
@@ -138,9 +182,55 @@ func getContextKey(ctx context.Context, n name) key {
 	return key
 }
 
+func getCounterSet(ctx context.Context) *ptCounterSet {
+	if id := ctx.Value(counterSetKey); id != nil {
+		return id.(*ptCounterSet)
+	}
+	// It's not set anywhere and wasn't hoisted, so create it.
+	if bctx, ok := ctx.(*beamCtx); ok {
+		bctx.bs.mu.Lock()
+		cs := &ptCounterSet{
+			counters:      make(map[nameHash]*counter),
+			distributions: make(map[nameHash]*distribution),
+			gauges:        make(map[nameHash]*gauge),
+		}
+		bctx.bs.css = append(bctx.bs.css, cs)
+		bctx.cs = cs
+		bctx.bs.mu.Unlock()
+		return cs
+	}
+	panic("counterSet missing, beam isn't set up properly.")
+	return nil // never runs.
+}
+
+type kind uint8
+
+const (
+	kindUnknown kind = iota
+	kindSumCounter
+	kindDistribution
+	kindGauge
+)
+
+func (t kind) String() string {
+	switch t {
+	case kindSumCounter:
+		return "Counter"
+	case kindDistribution:
+		return "Distribution"
+	case kindGauge:
+		return "Gauge"
+	default:
+		panic(fmt.Sprintf("Unknown metric type value: %v", uint8(t)))
+	}
+}
+
 // userMetric knows how to convert it's value to a Metrics_User proto.
+// TODO(lostluck): Move proto translation to the harness package to
+// avoid the proto dependency outside of the harness.
 type userMetric interface {
 	toProto() *fnexecution_v1.Metrics_User
+	kind() kind
 }
 
 // name is a pair of strings identifying a specific metric.
@@ -159,43 +249,87 @@ func newName(ns, n string) name {
 	return name{namespace: ns, name: n}
 }
 
+// We hash the name to a uint64 so we avoid using go's native string hashing for
+// every use of a metrics. uint64s have faster lookup than strings as a result.
+// Collisions are possible, but statistically unlikely as namespaces and names
+// are usually short enough to avoid this.
+var (
+	hasherMu sync.Mutex
+	hasher   = fnv.New64a()
+)
+
+func hashName(ns, n string) nameHash {
+	hasherMu.Lock()
+	hasher.Reset()
+	var buf [64]byte
+	b := buf[:]
+	hashString(ns, b)
+	hashString(n, b)
+	h := hasher.Sum64()
+	hasherMu.Unlock()
+	return nameHash(h)
+}
+
+// hashString hashes a string with the package level hasher
+// and requires posession of the hasherMu lock. The byte
+// slice is assumed to be backed by a [64]byte.
+func hashString(s string, b []byte) {
+	l := len(s)
+	i := 0
+	for len(s)-i > 64 {
+		n := i + 64
+		copy(b, s[i:n])
+		ioutilx.WriteUnsafe(hasher, b)
+		i = n
+	}
+	n := l - i
+	copy(b, s[i:])
+	ioutilx.WriteUnsafe(hasher, b[:n])
+}
+
 type key struct {
 	name               name
 	bundle, ptransform string
 }
 
 var (
-	// mu protects access to store
+	// mu protects access to the global store
 	mu sync.RWMutex
 	// store is a map of BundleIDs to PtransformIDs to userMetrics.
 	// it permits us to extract metric protos for runners per data Bundle, and
 	// per PTransform.
+	// TODO(lostluck): Migrate exec/plan.go to manage it's own perBundle counter stores.
+	// Note: This is safe to use to read the metrics concurrently with user modification
+	// in bundles, as only initial use modifies this map, and only contains the resulting
+	// metrics.
 	store = make(map[string]map[string]map[name]userMetric)
-
-	// We store the user path access to the cells in metric type segregated
-	// sync.Maps. Using sync.Maps lets metrics with disjoint keys have concurrent
-	// access to the cells, and using separate sync.Map per metric type
-	// simplifies code understanding, since each only contains a single type of
-	// cell.
-
-	countersMu      sync.RWMutex
-	counters        = make(map[key]*counter)
-	distributionsMu sync.RWMutex
-	distributions   = make(map[key]*distribution)
-	gaugesMu        sync.RWMutex
-	gauges          = make(map[key]*gauge)
 )
 
-// TODO(lostluck): 2018/03/05 Use a common internal beam now() instead, once that exists.
-var now = time.Now
+// storeMetric stores a metric away on its first use so it may be retrieved later on.
+// In the event of a name collision, storeMetric can panic, so it's prudent to release
+// locks if they are no longer required.
+func storeMetric(key key, m userMetric) {
+	mu.Lock()
+	defer mu.Unlock()
+	if _, ok := store[key.bundle]; !ok {
+		store[key.bundle] = make(map[string]map[name]userMetric)
+	}
+	if _, ok := store[key.bundle][key.ptransform]; !ok {
+		store[key.bundle][key.ptransform] = make(map[name]userMetric)
+	}
+	if ms, ok := store[key.bundle][key.ptransform][key.name]; ok {
+		if ms.kind() != m.kind() {
+			panic(fmt.Sprintf("metric name %s being reused for a second metric in a single PTransform", key.name))
+		}
+		return
+	}
+	store[key.bundle][key.ptransform][key.name] = m
+}
 
 // Counter is a simple counter for incrementing and decrementing a value.
 type Counter struct {
 	name name
-	// The following are a fast cache of the key and storage
-	mu sync.Mutex
-	k  key
-	c  *counter
+	hash nameHash
 }
 
 func (m *Counter) String() string {
@@ -204,38 +338,26 @@ func (m *Counter) String() string {
 
 // NewCounter returns the Counter with the given namespace and name.
 func NewCounter(ns, n string) *Counter {
-	mn := newName(ns, n)
 	return &Counter{
-		name: mn,
+		name: newName(ns, n),
+		hash: hashName(ns, n),
 	}
 }
 
 // Inc increments the counter within the given PTransform context by v.
 func (m *Counter) Inc(ctx context.Context, v int64) {
-	key := getContextKey(ctx, m.name)
-	cs := &counter{
-		value: v,
-	}
-	m.mu.Lock()
-	if m.k == key {
-		m.c.inc(v)
-		m.mu.Unlock()
-		return
-	}
-	m.k = key
-	countersMu.Lock()
-	if c, loaded := counters[key]; loaded {
-		m.c = c
-		countersMu.Unlock()
-		m.mu.Unlock()
+	cs := getCounterSet(ctx)
+	if c, ok := cs.counters[m.hash]; ok {
 		c.inc(v)
 		return
 	}
-	m.c = cs
-	counters[key] = cs
-	countersMu.Unlock()
-	m.mu.Unlock()
-	storeMetric(key, cs)
+	// We're the first to create this metric!
+	c := &counter{
+		value: v,
+	}
+	cs.counters[m.hash] = c
+	key := getContextKey(ctx, m.name)
+	storeMetric(key, c)
 }
 
 // Dec decrements the counter within the given PTransform context by v.
@@ -246,13 +368,10 @@ func (m *Counter) Dec(ctx context.Context, v int64) {
 // counter is a metric cell for counter values.
 type counter struct {
 	value int64
-	mu    sync.Mutex
 }
 
 func (m *counter) inc(v int64) {
-	m.mu.Lock()
-	m.value += v
-	m.mu.Unlock()
+	atomic.AddInt64(&m.value, v)
 }
 
 func (m *counter) String() string {
@@ -262,25 +381,23 @@ func (m *counter) String() string {
 // toProto returns a Metrics_User populated with the Data messages, but not the name. The
 // caller needs to populate with the metric's name.
 func (m *counter) toProto() *fnexecution_v1.Metrics_User {
-	m.mu.Lock()
-	defer m.mu.Unlock()
 	return &fnexecution_v1.Metrics_User{
 		Data: &fnexecution_v1.Metrics_User_CounterData_{
 			CounterData: &fnexecution_v1.Metrics_User_CounterData{
-				Value: m.value,
+				Value: atomic.LoadInt64(&m.value),
 			},
 		},
 	}
 }
 
+func (m *counter) kind() kind {
+	return kindSumCounter
+}
+
 // Distribution is a simple distribution of values.
 type Distribution struct {
 	name name
-
-	// The following are a fast cache of the key and storage
-	mu sync.Mutex
-	k  key
-	d  *distribution
+	hash nameHash
 }
 
 func (m *Distribution) String() string {
@@ -289,41 +406,29 @@ func (m *Distribution) String() string {
 
 // NewDistribution returns the Distribution with the given namespace and name.
 func NewDistribution(ns, n string) *Distribution {
-	mn := newName(ns, n)
 	return &Distribution{
-		name: mn,
+		name: newName(ns, n),
+		hash: hashName(ns, n),
 	}
 }
 
 // Update updates the distribution within the given PTransform context with v.
 func (m *Distribution) Update(ctx context.Context, v int64) {
-	key := getContextKey(ctx, m.name)
-	ds := &distribution{
+	cs := getCounterSet(ctx)
+	if d, ok := cs.distributions[m.hash]; ok {
+		d.update(v)
+		return
+	}
+	// We're the first to create this metric!
+	d := &distribution{
 		count: 1,
 		sum:   v,
 		min:   v,
 		max:   v,
 	}
-	m.mu.Lock()
-	if m.k == key {
-		m.d.update(v)
-		m.mu.Unlock()
-		return
-	}
-	m.k = key
-	distributionsMu.Lock()
-	if d, loaded := distributions[key]; loaded {
-		m.d = d
-		distributionsMu.Unlock()
-		m.mu.Unlock()
-		d.update(v)
-		return
-	}
-	m.d = ds
-	distributions[key] = ds
-	distributionsMu.Unlock()
-	m.mu.Unlock()
-	storeMetric(key, ds)
+	cs.distributions[m.hash] = d
+	key := getContextKey(ctx, m.name)
+	storeMetric(key, d)
 }
 
 // distribution is a metric cell for distribution values.
@@ -366,14 +471,14 @@ func (m *distribution) toProto() *fnexecution_v1.Metrics_User {
 	}
 }
 
+func (m *distribution) kind() kind {
+	return kindDistribution
+}
+
 // Gauge is a time, value pair metric.
 type Gauge struct {
 	name name
-
-	// The following are a fast cache of the key and storage
-	mu sync.Mutex
-	k  key
-	g  *gauge
+	hash nameHash
 }
 
 func (m *Gauge) String() string {
@@ -382,57 +487,30 @@ func (m *Gauge) String() string {
 
 // NewGauge returns the Gauge with the given namespace and name.
 func NewGauge(ns, n string) *Gauge {
-	mn := newName(ns, n)
 	return &Gauge{
-		name: mn,
+		name: newName(ns, n),
+		hash: hashName(ns, n),
 	}
 }
+
+// TODO(lostluck): 2018/03/05 Use a common internal beam now() instead, once that exists.
+var now = time.Now
 
 // Set sets the gauge to the given value, and associates it with the current time on the clock.
 func (m *Gauge) Set(ctx context.Context, v int64) {
-	key := getContextKey(ctx, m.name)
-	gs := &gauge{
-		t: now(),
-		v: v,
-	}
-	m.mu.Lock()
-	if m.k == key {
-		m.g.set(v)
-		m.mu.Unlock()
-		return
-	}
-	m.k = key
-	gaugesMu.Lock()
-	if g, loaded := gauges[key]; loaded {
-		m.g = g
-		gaugesMu.Unlock()
-		m.mu.Unlock()
+	cs := getCounterSet(ctx)
+	if g, ok := cs.gauges[m.hash]; ok {
 		g.set(v)
 		return
 	}
-	m.g = gs
-	gauges[key] = gs
-	gaugesMu.Unlock()
-	m.mu.Unlock()
-	storeMetric(key, gs)
-}
-
-// storeMetric stores a metric away on its first use so it may be retrieved later on.
-// In the event of a name collision, storeMetric can panic, so it's prudent to release
-// locks if they are no longer required.
-func storeMetric(key key, m userMetric) {
-	mu.Lock()
-	defer mu.Unlock()
-	if _, ok := store[key.bundle]; !ok {
-		store[key.bundle] = make(map[string]map[name]userMetric)
+	// We're the first to create this metric!
+	g := &gauge{
+		t: now(),
+		v: v,
 	}
-	if _, ok := store[key.bundle][key.ptransform]; !ok {
-		store[key.bundle][key.ptransform] = make(map[name]userMetric)
-	}
-	if _, ok := store[key.bundle][key.ptransform][key.name]; ok {
-		panic(fmt.Sprintf("metric name %s being reused for a second metric in a single PTransform", key.name))
-	}
-	store[key.bundle][key.ptransform][key.name] = m
+	cs.gauges[m.hash] = g
+	key := getContextKey(ctx, m.name)
+	storeMetric(key, g)
 }
 
 // gauge is a metric cell for gauge values.
@@ -466,8 +544,12 @@ func (m *gauge) toProto() *fnexecution_v1.Metrics_User {
 	}
 }
 
+func (m *gauge) kind() kind {
+	return kindGauge
+}
+
 func (m *gauge) String() string {
-	return fmt.Sprintf("time: %s value: %d", m.t, m.v)
+	return fmt.Sprintf("%v time: %s value: %d", m.kind(), m.t, m.v)
 }
 
 // ToProto exports all collected metrics for the given BundleID and PTransform ID pair.
@@ -505,23 +587,19 @@ func DumpToOut() {
 func dumpTo(p func(format string, args ...interface{})) {
 	mu.RLock()
 	defer mu.RUnlock()
-	countersMu.RLock()
-	defer countersMu.RUnlock()
-	distributionsMu.RLock()
-	defer distributionsMu.RUnlock()
-	gaugesMu.RLock()
-	defer gaugesMu.RUnlock()
 	var bs []string
 	for b := range store {
 		bs = append(bs, b)
 	}
 	sort.Strings(bs)
+
 	for _, b := range bs {
 		var pts []string
 		for pt := range store[b] {
 			pts = append(pts, pt)
 		}
 		sort.Strings(pts)
+
 		for _, pt := range pts {
 			var ns []name
 			for n := range store[b][pt] {
@@ -537,17 +615,10 @@ func dumpTo(p func(format string, args ...interface{})) {
 				return false
 			})
 			p("Bundle: %q - PTransformID: %q", b, pt)
+
 			for _, n := range ns {
-				key := key{name: n, bundle: b, ptransform: pt}
-				if m, ok := counters[key]; ok {
-					p("\t%s - %s", key.name, m)
-				}
-				if m, ok := distributions[key]; ok {
-					p("\t%s - %s", key.name, m)
-				}
-				if m, ok := gauges[key]; ok {
-					p("\t%s - %s", key.name, m)
-				}
+				m := store[b][pt][n]
+				p("\t%s - %s", n, m)
 			}
 		}
 	}
@@ -558,9 +629,6 @@ func dumpTo(p func(format string, args ...interface{})) {
 func Clear() {
 	mu.Lock()
 	store = make(map[string]map[string]map[name]userMetric)
-	counters = make(map[key]*counter)
-	distributions = make(map[key]*distribution)
-	gauges = make(map[key]*gauge)
 	mu.Unlock()
 }
 
@@ -570,21 +638,6 @@ func ClearBundleData(b string) {
 	// No concurrency races since mu guards all access to store,
 	// and the metric cell sync.Maps are goroutine safe.
 	mu.Lock()
-	pts := store[b]
-	countersMu.Lock()
-	distributionsMu.Lock()
-	gaugesMu.Lock()
-	for pt, m := range pts {
-		for n := range m {
-			key := key{name: n, bundle: b, ptransform: pt}
-			delete(counters, key)
-			delete(distributions, key)
-			delete(gauges, key)
-		}
-	}
-	countersMu.Unlock()
-	distributionsMu.Unlock()
-	gaugesMu.Unlock()
 	delete(store, b)
 	mu.Unlock()
 }