… sdks/go/pkg/beam/core/metrics/metrics_test.go