@lostluck 2020-01-23
So per the beam model, Bundles are the unit of a reproducible computation. If the bundle needs to be re-tried, the runner needs to be able to throw away the metrics for the bad bundle, so the final counters can be correct. As such, they are "logical" counters, rather than "physical" counters.

It's the runners job to aggregate metrics for all bundles, for final presentation, since any given worker isn't aware of the full distributed value of a counter.

To answer your questions: Yes, each bundle has their own instances of counters. These metrics get reported to the runner, and are currently "extracted" from the metric cells via the ToProto function (which requires the bundleID and PTransformID in question). It's the up to the runner to provide a way for the user to see them.
There's unfinished work on the "Job Management" proto to provide a common portable way for SDKs to get metrics from runners. Because of this, the Go SDK doesn't provide programmatic access to metrics after a job, so it's up to the user to access the data from the runner directly.